requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-language==2.13.3
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
requests==2.31.0
rsa==4.9
urllib3==2.2.1
Werkzeug==3.0.2
